[backend] feat: Improve document deletion and add related tests - Add null check for isDocumentOwner to return 404 when document is not found. - Create new tests for document deletion, including checks for owner status and existence of document.
[terraform] chore: Add .gitignore
- Add a .gitignore file in the terraform directory to ignore state files and backups.

+++ backend/controllers/documents.js
@@ -68,10 +68,14 @@ export const deleteDocument = async (req, res) => {
 
   // auth the user
   const isOwner = await isDocumentOwner(req.user, document_id);
+  if (isOwner === null)
+    return res.status(404).json({ error: "Document not found" });
+
   if (!isOwner)
     return res.status(403).json({ error: "forbidden (not the owner)" });
 
   const document = await deleteDoc(document_id);
+  console.log({ document });
   if (!document) return res.status(404).json({ error: "Document not found" });
 
   return res
